Adds replay buffer files and docs
