+++ chainerrl/replay_buffers/__init__.py
@@ -1,0 +1,5 @@
+from chainerrl.replay_buffers.episodic import EpisodicReplayBuffer  # NOQA
+from chainerrl.replay_buffers.prioritized import PrioritizedReplayBuffer  # NOQA
+from chainerrl.replay_buffers.prioritized import PriorityWeightError  # NOQA
+from chainerrl.replay_buffers.prioritized_episodic import PrioritizedEpisodicReplayBuffer  # NOQA
+from chainerrl.replay_buffers.replay_buffer import ReplayBuffer  # NOQA
