+++ docs/replay_buffers.rst
@@ -1,0 +1,21 @@
+======
+Replay Buffers
+======
+
+ReplayBuffer interfaces
+================
+
+.. autoclass:: chainerrl.replay_buffer.ReplayBuffer
+   :members:
+
+Agent implementations
+=====================
+
+.. autoclass:: chainerrl.replay_buffer.EpisodicReplayBuffer
+
+.. autoclass:: chainerrl.replay_buffer.ReplayBuffer
+
+.. autoclass:: chainerrl.replay_buffer.PrioritizedReplayBuffer
+
+.. autoclass:: chainerrl.replay_buffer.PrioritizedEpisodicReplayBuffer
+
